refactor(api): wip fixing files
