--- packages/api/src/controllers/tasks/index.ts
@@ -1,8 +1,0 @@
-export * from "./createTask";
-export * from "./getTask";
-export * from "./deleteTask";
-export * from "./editTask";
-export * from "./comments/createComment";
-export * from "./comments/deleteComment";
-export * from "./comments/editComment";
-export * from "./comments/getComment";
