--- packages/api/src/models/index.ts
@@ -1,13 +1,0 @@
-import Tasks from "./Task";
-import Users from "./User";
-import ConfirmationToken from "./ConfirmationToken";
-import PasswordResetToken from "./PasswordResetToken";
-import Workspaces from "../components/workspaces/model";
-
-export default {
-  Tasks,
-  Users,
-  Workspaces,
-  ConfirmationToken,
-  PasswordResetToken,
-};
